% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_RL_MH_FUN_2.R
\name{A_samp_fun}
\alias{A_samp_fun}
\title{A_samp_fun}
\usage{
A_samp_fun(
  R,
  W_samp,
  G_samp,
  B_samp,
  motif_len_w,
  motif_len_g,
  dict,
  theta_0_samp,
  theta_samp,
  theta_til_samp
)
}
\arguments{
\item{R}{A vector of sequences.}

\item{W_samp}{A vector of sampled W labels.}

\item{G_samp}{A vector of sampled G labels.}

\item{B_samp}{A vector of sampled positions for the second motif.}

\item{motif_len_w}{An integer specifying the length of the first binding motif.}

\item{motif_len_g}{An integer specifying the length of the second binding motif.}

\item{dict}{A dictionary of possible motif elements.}

\item{theta_0_samp}{A vector of sampled background probabilities.}

\item{theta_samp}{A matrix of sampled probabilities for the first motif.}

\item{theta_til_samp}{A matrix of sampled probabilities for the second motif.}
}
\value{
A list containing the sampled positions of the first binding motif.
}
\description{
A function to sample the position of the first binding motif using posterior probabilities.
}
