% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RL_MH_A_THETA_W_FUN_2.R
\name{A_theta_samp_fun}
\alias{A_theta_samp_fun}
\title{A_theta_samp_fun}
\usage{
A_theta_samp_fun(
  R,
  W_samp,
  G_samp,
  UW_loc,
  A_samp,
  B_samp,
  motif_len_w,
  motif_len_g,
  dict,
  theta_0_samp,
  theta_samp,
  theta_til_samp,
  pri_alw,
  pri_Wp
)
}
\arguments{
\item{R}{A vector representing the observed sequences.}

\item{W_samp}{A vector indicating the presence of the first motif (sampled).}

\item{G_samp}{A vector indicating the presence of the second motif (sampled).}

\item{UW_loc}{Locations where \code{W} is unknown.}

\item{A_samp}{A vector of positions for the first motif (sampled).}

\item{B_samp}{A vector of positions for the second motif (sampled).}

\item{motif_len_w}{Length of the first motif.}

\item{motif_len_g}{Length of the second motif.}

\item{dict}{A dictionary of possible characters in the sequences.}

\item{theta_0_samp}{Background probabilities for each character.}

\item{theta_samp}{Probabilities for the first motif (sampled).}

\item{theta_til_samp}{Probabilities for the second motif (sampled).}

\item{pri_alw}{Prior Dirichlet parameter for \code{theta}.}

\item{pri_Wp}{Prior probability for \code{W}.}
}
\value{
A list containing updated values for \code{A}, \code{theta}, and \code{W}, along with the jump indicator.
}
\description{
This function performs Metropolis-Hastings sampling for the positions of the first motif (\code{A}) and its probabilities (\code{theta}).
}
