% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demotif_RL_MH_FUN.R
\name{de_motif}
\alias{de_motif}
\title{de_motif}
\usage{
de_motif(
  Data,
  motif_len_w,
  motif_len_g,
  mh_step_w,
  mh_step_g,
  stop_jump_step,
  N
)
}
\arguments{
\item{Data}{A data frame containing the sequence, the label vector W, and the label vector G.}

\item{motif_len_w}{An integer specifying the length of the first binding motif.}

\item{motif_len_g}{An integer specifying the length of the second binding motif.}

\item{mh_step_w}{An integer specifying the number of steps to skip between each execution of the MH algorithm for the first binding motif.}

\item{mh_step_g}{An integer specifying the number of steps to skip between each execution of the MH algorithm for the second binding motif.}

\item{stop_jump_step}{An integer specifying the time step at which to stop the MH algorithm for the first and second binding motifs.}

\item{N}{The number of sampling iterations to perform.}
}
\value{
NULL. Results are saved to files in specified directories.
}
\description{
A function to perform motif detection with Metropolis-Hastings (MH) sampling.
}
