% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_MH_FUN_2.R
\name{g_samp_fun}
\alias{g_samp_fun}
\title{g_samp_fun}
\usage{
g_samp_fun(
  R,
  W_samp,
  UG_loc,
  A_samp,
  B_samp,
  motif_len_w,
  motif_len_g,
  dict,
  theta_0_samp,
  theta_samp,
  theta_til_samp,
  pri_gp
)
}
\arguments{
\item{R}{A vector containing sequence data.}

\item{W_samp}{A vector containing sampled W values.}

\item{UG_loc}{A vector of indices where G is unobserved (NA values).}

\item{A_samp}{A vector containing sampled A values.}

\item{B_samp}{A vector containing sampled B values.}

\item{motif_len_w}{The length of the first motif.}

\item{motif_len_g}{The length of the second motif.}

\item{dict}{A dictionary containing unique characters from the sequences.}

\item{theta_0_samp}{Sampled values for theta_0.}

\item{theta_samp}{Sampled values for theta.}

\item{theta_til_samp}{Sampled values for theta_til.}

\item{pri_gp}{The prior probability of G being 1.}
}
\value{
A list containing:
- \code{G_samp_unobs}: The sampled values for unobserved G.
- \code{Post_prob}: A matrix of posterior probabilities for each unobserved G.
}
\description{
g_samp_fun
}
