% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/THETA_0_FUN_2.R
\name{h_00}
\alias{h_00}
\title{h_00}
\usage{
h_00(Ri, dict)
}
\arguments{
\item{Ri}{A string representing a sequence of characters.}

\item{dict}{A dictionary of possible characters in the sequences.}
}
\value{
A vector representing the count of each character in \code{Ri}.
}
\description{
h_00
}
