% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/THETA_0_FUN_2.R
\name{h_01}
\alias{h_01}
\title{h_01}
\usage{
h_01(Ri, bi, dict)
}
\arguments{
\item{Ri}{A string representing a sequence of characters.}

\item{bi}{A vector of positions to exclude from \code{Ri}.}

\item{dict}{A dictionary of possible characters in the sequences.}
}
\value{
A vector representing the count of each character in \code{Ri} excluding positions in \code{bi}.
}
\description{
h_01
}
