% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/THETA_0_FUN_2.R
\name{h_10}
\alias{h_10}
\title{h_10}
\usage{
h_10(Ri, ai, dict)
}
\arguments{
\item{Ri}{A string representing a sequence of characters.}

\item{ai}{A vector of positions to exclude from \code{Ri}.}

\item{dict}{A dictionary of possible characters in the sequences.}
}
\value{
A vector representing the count of each character in \code{Ri} excluding positions in \code{ai}.
}
\description{
h_10
}
