% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOGLLK_FUN_2.R
\name{logllk_fun}
\alias{logllk_fun}
\title{logllk_fun}
\usage{
logllk_fun(
  R,
  W_samp,
  G_samp,
  A_samp,
  B_samp,
  motif_len_w,
  motif_len_g,
  dict,
  theta_0_samp,
  theta_samp,
  theta_til_samp
)
}
\arguments{
\item{R}{A vector of sequences.}

\item{W_samp}{A vector indicating the presence/absence of the first motif (W).}

\item{G_samp}{A vector indicating the presence/absence of the second motif (G).}

\item{A_samp}{A vector of positions for the first motif in each sequence.}

\item{B_samp}{A vector of positions for the second motif in each sequence.}

\item{motif_len_w}{The length of the first motif.}

\item{motif_len_g}{The length of the second motif.}

\item{dict}{A dictionary of unique characters in the sequences.}

\item{theta_0_samp}{A vector of sampled probabilities for the background model.}

\item{theta_samp}{A matrix of sampled probabilities for the first motif.}

\item{theta_til_samp}{A matrix of sampled probabilities for the second motif.}
}
\value{
A numeric value representing the total log-likelihood across all sequences.
}
\description{
This function calculates the total log-likelihood for a set of sequences based on the sampled motifs.
}
