% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOGLLK_FUN_2.R
\name{logllk_fun_00}
\alias{logllk_fun_00}
\title{logllk_fun_00}
\usage{
logllk_fun_00(Ri, theta_0_samp, dict)
}
\arguments{
\item{Ri}{A string representing the sequence.}

\item{theta_0_samp}{A vector of sampled probabilities for the background model.}

\item{dict}{A dictionary of unique characters in the sequence.}
}
\value{
A numeric value representing the log-likelihood of the sequence.
}
\description{
This function calculates the log-likelihood of a sequence when both motifs are absent (wi = 0, gi = 0).
}
