% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_B_RL_MH_FUN.R
\name{pred_B_samp_fun}
\alias{pred_B_samp_fun}
\title{pred_B_samp_fun}
\usage{
pred_B_samp_fun(
  R,
  W_samp,
  G_samp,
  A_samp,
  motif_len_w,
  motif_len_g,
  dict,
  theta_0_samp,
  theta_samp,
  theta_til_samp
)
}
\arguments{
\item{R}{A sequence.}

\item{W_samp}{A sampled W label.}

\item{G_samp}{A sampled G label.}

\item{A_samp}{A sampled position for the first motif.}

\item{motif_len_w}{An integer specifying the length of the first binding motif.}

\item{motif_len_g}{An integer specifying the length of the second binding motif.}

\item{dict}{A dictionary of possible motif elements.}

\item{theta_0_samp}{A vector of sampled background probabilities.}

\item{theta_samp}{A matrix of sampled probabilities for the first motif.}

\item{theta_til_samp}{A matrix of sampled probabilities for the second motif.}
}
\value{
A list containing the sampled positions of the second binding motif.
}
\description{
A function to sample the position of the second binding motif using posterior probabilities.
}
