% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RL_MH_B_THETIL_G_FUN_2.R
\name{prob_g_fun}
\alias{prob_g_fun}
\title{prob_g_fun}
\usage{
prob_g_fun(
  R,
  W_samp,
  G,
  UG_loc,
  A_samp,
  B,
  dict,
  theta_0_samp,
  theta_samp,
  theta_til,
  pri_gp
)
}
\arguments{
\item{R}{A vector representing the observed sequences.}

\item{W_samp}{A vector indicating the presence of the first motif (sampled).}

\item{G}{A vector indicating the presence of the second motif.}

\item{UG_loc}{Locations where \code{G} is unknown.}

\item{A_samp}{A vector of positions for the first motif (sampled).}

\item{B}{A vector of positions for the second motif.}

\item{dict}{A dictionary of possible characters in the sequences.}

\item{theta_0_samp}{Background probabilities for each character.}

\item{theta_samp}{Probabilities for the first motif (sampled).}

\item{theta_til}{Probabilities for the second motif.}

\item{pri_gp}{Prior probability for \code{G}.}
}
\value{
A numeric value representing the posterior probability of \code{G}.
}
\description{
This function calculates the posterior probability of \code{G} (presence of the second motif).
}
