% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RL_MH_A_THETA_W_FUN_2.R
\name{prob_theta_fun}
\alias{prob_theta_fun}
\title{prob_theta_fun}
\usage{
prob_theta_fun(R, W, G_samp, A, B_samp, dict, theta, pri_alw)
}
\arguments{
\item{R}{A vector representing the observed sequences.}

\item{W}{A vector indicating the presence of the first motif.}

\item{G_samp}{A vector indicating the presence of the second motif.}

\item{A}{A vector of positions for the first motif.}

\item{B_samp}{A vector of positions for the second motif.}

\item{dict}{A dictionary of possible characters in the sequences.}

\item{theta}{Probabilities for the first motif.}

\item{pri_alw}{Prior Dirichlet parameter for \code{theta}.}
}
\value{
A numeric value representing the posterior probability of \code{theta}.
}
\description{
This function calculates the posterior probability of \code{theta} based on the observed sequences and other parameters.
}
