% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RL_MH_B_THETIL_G_FUN_2.R
\name{prob_theta_til_fun}
\alias{prob_theta_til_fun}
\title{prob_theta_til_fun}
\usage{
prob_theta_til_fun(R, W_samp, G, A_samp, B, dict, theta_til, pri_alg)
}
\arguments{
\item{R}{A vector representing the observed sequences.}

\item{W_samp}{A vector indicating the presence of the first motif (sampled).}

\item{G}{A vector indicating the presence of the second motif.}

\item{A_samp}{A vector of positions for the first motif (sampled).}

\item{B}{A vector of positions for the second motif.}

\item{dict}{A dictionary of possible characters in the sequences.}

\item{theta_til}{Probabilities for the second motif.}

\item{pri_alg}{Prior Dirichlet parameter for \code{theta_til}.}
}
\value{
A numeric value representing the posterior probability of \code{theta_til}.
}
\description{
This function calculates the posterior probability of \code{theta_til} based on the observed sequences and other parameters.
}
