% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RL_MH_A_THETA_W_FUN_2.R
\name{prob_w_fun}
\alias{prob_w_fun}
\title{prob_w_fun}
\usage{
prob_w_fun(
  R,
  W,
  G_samp,
  UW_loc,
  A,
  B_samp,
  dict,
  theta_0_samp,
  theta,
  theta_til_samp,
  pri_wp
)
}
\arguments{
\item{R}{A vector representing the observed sequences.}

\item{W}{A vector indicating the presence of the first motif.}

\item{G_samp}{A vector indicating the presence of the second motif.}

\item{UW_loc}{Locations where \code{W} is unknown.}

\item{A}{A vector of positions for the first motif.}

\item{B_samp}{A vector of positions for the second motif.}

\item{dict}{A dictionary of possible characters in the sequences.}

\item{theta_0_samp}{Background probabilities for each character.}

\item{theta}{Probabilities for the first motif.}

\item{theta_til_samp}{Probabilities for the second motif.}

\item{pri_wp}{Prior probability for \code{W}.}
}
\value{
A numeric value representing the posterior probability of \code{W}.
}
\description{
This function calculates the posterior probability of \code{W} (presence of the first motif).
}
