% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/THETA_0_FUN_2.R
\name{theta_0_samp_fun}
\alias{theta_0_samp_fun}
\title{theta_0_samp_fun}
\usage{
theta_0_samp_fun(
  R,
  W_samp,
  G_samp,
  A_samp,
  B_samp,
  motif_len_w,
  motif_len_g,
  dict,
  pri_al0
)
}
\arguments{
\item{R}{A vector of string sequences.}

\item{W_samp}{A vector indicating the presence of the first motif (sampled).}

\item{G_samp}{A vector indicating the presence of the second motif (sampled).}

\item{A_samp}{A vector of positions for the first motif (sampled).}

\item{B_samp}{A vector of positions for the second motif (sampled).}

\item{motif_len_w}{Length of the first motif.}

\item{motif_len_g}{Length of the second motif.}

\item{dict}{A dictionary of possible characters in the sequences.}

\item{pri_al0}{A prior Dirichlet parameter for the background probabilities.}
}
\value{
A vector representing the sampled background probabilities (\code{theta_0}).
}
\description{
theta_0_samp_fun
}
