% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/W_MH_FUN_2.R
\name{w_samp_fun}
\alias{w_samp_fun}
\title{w_samp_fun}
\usage{
w_samp_fun(
  R,
  G_samp,
  UW_loc,
  A_samp,
  B_samp,
  motif_len_w,
  motif_len_g,
  dict,
  theta_0_samp,
  theta_samp,
  theta_til_samp,
  pri_wp
)
}
\arguments{
\item{R}{A vector of string sequences.}

\item{G_samp}{A vector indicating the presence of the second motif (sampled).}

\item{UW_loc}{Indices of sequences with unknown \code{W} values (unobserved presence of the first motif).}

\item{A_samp}{A vector of positions for the first motif (sampled).}

\item{B_samp}{A vector of positions for the second motif (sampled).}

\item{motif_len_w}{Length of the first motif.}

\item{motif_len_g}{Length of the second motif.}

\item{dict}{A dictionary of possible characters in the sequences.}

\item{theta_0_samp}{Background probabilities.}

\item{theta_samp}{Probabilities for the first motif.}

\item{theta_til_samp}{Probabilities for the second motif.}

\item{pri_wp}{Prior probability of the presence of the first motif (\code{W = 1}).}
}
\value{
A list containing:
\itemize{
\item \code{W_samp_unobs}: A vector of sampled values for the unobserved \code{W} (1 or 0).
\item \code{Post_prob}: A matrix of posterior probabilities for \code{W = 1} and \code{W = 0} for each sequence in \code{UW_loc}.
}
}
\description{
w_samp_fun
}
